Move agentic judge framework to agent-mode-evals; add on_final_exchange_text and validate_agent_mode_llm_step hooks
- Remove agent_judge.rs, rubrics/ from llm_judge/: all eval-specific
  types, prompt building, rubric definitions, and scenario configs belong
  in agent-mode-evals, not in the warp codebase.
- Add on_final_exchange_text(f): thin hook that reads the final exchange
  text and passes it to a caller-supplied callback, so agent-mode-evals
  can inspect the judge's last message without touching BlocklistAIHistoryModel.
- Add validate_agent_mode_llm_step(model): wraps LLMPreferences access
  to validate a judge model ID, avoiding a panic on misconfigured models.
- Remove assert_rubric_from_final_message* from assertions.rs; the rubric
  assertion logic now lives alongside the rubric types in agent-mode-evals.
