+++ app/src/integration_testing/agent_mode/llm_judge/mod.rs
@@ -1,6 +1,4 @@
-pub mod agent_judge;
 pub mod llm_generate;
-pub mod rubrics;
 
 use anyhow::Result;
 use llm_generate::LLMGenerateRequest;
